+++ src/index.ts
@@ -6,16 +6,23 @@ export * from './structs/types.js';
 export * from './structs/utilities.js';
 export type {
   AnyStruct,
+  Assign,
   EnumSchema,
+  If,
   InferStructTuple,
   IsExactMatch,
   IsMatch,
   IsRecord,
   IsTuple,
+  IsUnion,
   ObjectSchema,
+  ObjectType,
   OmitBy,
   Optionalize,
+  PartialObjectSchema,
   PickBy,
   Simplify,
+  StructSchema,
+  TupleSchema,
   UnionToIntersection,
 } from './utils.js';